requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
shortuuid==1.0.13
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
